a pipeline version of user memory (exmem)
Note: In order not to have memory data access become a bottleneck, pipeline memory is needed ( exmem_pipeline.v). The code is written for the purpose to present the design structure. There is no syntax check nor simulation verification.  You can use it to replace the 10T non-pipeline user RAM.